[MCP]: add draft example descriptions for component api response
