+++ .vscode/launch.json
@@ -5,6 +5,12 @@
   "version": "0.2.0",
   "configurations": [
     {
+        "command": "yarn start",
+        "name": "Run yarn start",
+        "request": "launch",
+        "type": "node-terminal"
+    },
+    {
       "type": "node",
       "request": "launch",
       "name": "Jest",
